ME-2744: connector install flow should write token to config file (#345)
invite code based connector install flow, particularly used by the
one-liner docker run command that starts the connector with an invite
code, should also write connector token to the border0 config file,
which is located at `$HOME/.border0/config.yml`.

previously, in the connector install flow, connector token was only
written to files with invite code as file names, for example,
`$HOME/.border0/invite_<invite_code>`. with that, to restart the
same container, user has to run the same command with the exactly the
same invite code, but if the user is not aware of that, there is no way
to retrieve the invite code, and then get stuck and confused

with this new change, user can start the container again with or without
the invite code.

the invite code is convenient for starting up multiple docker containers
that run border0 connector, that said, most users likely only need to
run one connector with docker on a computer.